# teacher/backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TIDE Teacher API Contract
  version: 0.10.0
  description: >-
    教师端公共合同。包含教师端任务、文件和 My TIDE 安全读取；
    各模块实现状态见 contracts/README。任务、G01 状态和外部消息直接使用共享 PostgreSQL 契约表；
    教师积分与逐课积分直接读取世文当前聚合视图，不在教师端重算。
servers:
  - url: http://localhost:3000
    description: local backend
...
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: 使用公司邮箱和世文教师编号注册
      operationId: registerTeacher
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '202':
          description: 账号和绑定已创建，等待邮箱验证
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistrationAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/auth/email-verification/confirm:
    post:
      tags: [Auth]
      summary: 使用一次性令牌验证邮箱
      operationId: confirmEmailVerification
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmEmailRequest' }
      responses:
        '200':
          description: 邮箱已验证，重复使用已成功令牌时返回 ALREADY_VERIFIED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmailVerificationResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/auth/email-verification/resend:
    post:
      tags: [Auth]
      summary: 重新发送邮箱验证邮件
      description: 无论邮箱是否存在均返回相同响应，避免泄露账号状态。
      operationId: resendEmailVerification
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResendVerificationRequest' }
      responses:
        '202':
          description: 请求已接受
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResendAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 使用已验证邮箱和密码登录
      operationId: loginTeacher
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: 短期访问令牌和一次显示的刷新令牌
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokenPair' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { description: 尝试过于频繁 }
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: 轮换刷新令牌并签发新访问令牌
      operationId: refreshTeacherSession
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshSessionRequest' }
      responses:
        '200':
          description: 新令牌对；旧刷新令牌立即失效
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokenPair' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: 注销当前会话
      operationId: logoutTeacher
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: 当前会话已撤销 }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/password-reset/request:
    post:
      tags: [Auth]
      summary: 申请密码重置邮件
      description: 无论邮箱是否存在均返回相同响应，避免泄露账号状态。
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordResetRequest' }
      responses:
        '202':
          description: 请求已接受
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResendAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: 使用一次性令牌设置新密码
      operationId: confirmPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmPasswordResetRequest' }
      responses:
        '200':
          description: 密码已更新，所有既有会话已撤销
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PasswordResetCompleted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /api/v1/files/upload-intents:
    post:
      tags: [Files]
      summary: 创建受控文件上传意图
      operationId: createUploadIntent
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UploadIntentRequest' }
      responses:
        '201':
          description: 上传意图
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadIntentResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/files/{fileId}/content:
    put:
      tags: [Files]
      summary: 以流式请求向本地受控上传意图提交文件内容
      operationId: uploadFileContent
      security: [{ bearerAuth: [] }]
      parameters:
        - name: fileId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: 文件内容已接收，等待完成确认
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileObject' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { description: 文件超过环境配置的大小限制 }
        '415': { description: 文件 MIME 类型不在允许列表 }
    get:
      tags: [Files]
      summary: 鉴权下载当前老师已就绪的私有文件
      operationId: downloadFileContent
      security: [{ bearerAuth: [] }]
      parameters:
        - name: fileId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 私有文件内容
          headers:
            Cache-Control:
              schema: { type: string, const: 'private, no-store' }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/files/{fileId}/complete:
    post:
      tags: [Files]
      summary: 确认上传完成并触发服务端校验
      operationId: completeUpload
      security: [{ bearerAuth: [] }]
      parameters:
        - name: fileId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [sha256]
              properties:
                sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
      responses:
        '200':
          description: 文件状态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileObject' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/faq/conversations:
    post:
      tags: [FAQ]
      summary: 获取或创建当前老师的开放 FAQ 会话
      operationId: createFaqConversation
      security: [{ bearerAuth: [] }]
      responses:
        '201':
          description: 当前开放 FAQ 会话已返回；没有时创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaqConversationCreated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/faq/conversations/{conversationId}:
    get:
      tags: [FAQ]
      summary: 获取当前老师的 FAQ 会话和历史消息
      operationId: getFaqConversation
      security: [{ bearerAuth: [] }]
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 会话、消息和教师可见来源
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaqConversation' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/faq/conversations/{conversationId}/messages:
    post:
      tags: [FAQ]
      summary: 向 FAQ 会话发送纯文字问题
      description: 不接收图片或附件；先在已生效 FAQ 中做英文自然语义匹配，意图不明确时追问，确认未命中时拒绝猜测并记录 Gap。
      operationId: askFaqQuestion
      security: [{ bearerAuth: [] }]
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AskFaqRequest' }
      responses:
        '200':
          description: 保存后的教师消息和安全 FAQ 回答
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaqAnswerResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/faq/messages/{messageId}/feedback:
    post:
      tags: [FAQ]
      summary: 保存当前老师对 FAQ 回答的首次反馈
      operationId: saveFaqFeedback
      security: [{ bearerAuth: [] }]
      parameters:
        - name: messageId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FaqFeedbackRequest' }
      responses:
        '200':
          description: 反馈已保存；完全相同的重复请求幂等返回
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaqFeedbackResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/app-events:
    post:
      tags: [Analytics]
      summary: 保存已登录老师的白名单产品分析事件
      description: >-
        事件按 event_id 幂等保存。任务事件必须绑定当前老师拥有的 assignment；
        任务编码、类型、模板版本、执行规则版本、状态和尝试次数由后端权威数据覆盖。
        埋点失败不改变任务、积分或课程业务状态。当前接口独立限流为每分钟 300 次。
      operationId: createAppEvent
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAppEvent' }
      responses:
        '202':
          description: 事件已接受或幂等忽略
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppEventAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { description: properties 序列化后超过 8 KB }
        '429': { description: 埋点请求过于频繁；客户端按 Retry-After 重试 }
  /api/v1/app-events/anonymous:
    post:
      tags: [Analytics]
      summary: 保存登录前页面与技术异常事件
      description: >-
        只接受 PAGE_VIEWED、PAGE_LOAD_SUCCEEDED、PAGE_LOAD_FAILED、
        NAVIGATION_CLICKED、LANGUAGE_CHANGED、PAGE_EXITED、API_FAILED、
        NETWORK_TIMEOUT、FRONTEND_UNCAUGHT_ERROR，且不得携带 taskAssignmentId。
        当前接口独立限流为每分钟 30 次。
      operationId: createAnonymousAppEvent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAppEvent' }
      responses:
        '202':
          description: 匿名事件已接受或幂等忽略
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppEventAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '413': { description: properties 序列化后超过 8 KB }
        '429': { description: 匿名埋点请求过于频繁；客户端按 Retry-After 重试 }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
...
    Unauthorized:
      description: 未认证或验签失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    BadRequest:
      description: 请求字段格式错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: 资源不存在或不属于当前教师
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: 幂等负载冲突、状态版本冲突或非法状态迁移
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ValidationError:
      description: 请求字段或任务输出不符合合同
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ServiceUnavailable:
      description: 数据源或外部依赖暂时不可用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, teacherId, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        teacherId: { type: string, minLength: 1, maxLength: 128 }
        password: { type: string, minLength: 12, maxLength: 128 }
    RegistrationAccepted:
      type: object
      additionalProperties: false
      required: [status, verificationEmailSent]
      properties:
        status: { type: string, const: VERIFICATION_REQUIRED }
        verificationEmailSent: { type: boolean }
    ConfirmEmailRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: { type: string, minLength: 32, maxLength: 512 }
    EmailVerificationResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [VERIFIED, ALREADY_VERIFIED]
    ResendVerificationRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 254 }
    ResendAccepted:
      type: object
      additionalProperties: false
      required: [accepted]
      properties:
        accepted: { type: boolean, const: true }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 1, maxLength: 128 }
    RefreshSessionRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken: { type: string, minLength: 32, maxLength: 512 }
    AuthTokenPair:
      type: object
      additionalProperties: false
      required: [tokenType, accessToken, accessTokenExpiresIn, refreshToken, refreshTokenExpiresAt]
      properties:
        tokenType: { type: string, const: Bearer }
        accessToken: { type: string }
        accessTokenExpiresIn: { type: integer, minimum: 60 }
        refreshToken: { type: string }
        refreshTokenExpiresAt: { type: string, format: date-time }
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 254 }
    ConfirmPasswordResetRequest:
      type: object
      additionalProperties: false
      required: [token, newPassword]
      properties:
        token: { type: string, minLength: 32, maxLength: 512 }
        newPassword: { type: string, minLength: 12, maxLength: 128 }
    PasswordResetCompleted:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string, const: PASSWORD_UPDATED }
...
    UploadIntentRequest:
      type: object
      additionalProperties: false
      required: [taskInstanceId, stepKey, filename, mimeType, sizeBytes, sha256]
      properties:
        taskInstanceId: { type: string }
        stepKey: { type: string }
        filename: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer, minimum: 1 }
        sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
    UploadIntentResponse:
      type: object
      required: [fileId, uploadMethod, uploadUrl, expiresAt, requiredHeaders, requiredFields]
      properties:
        fileId: { type: string, format: uuid }
        uploadMethod: { type: string, enum: [LOCAL_STREAM, OSS_POST_FORM] }
        uploadUrl: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
        requiredHeaders: { type: object, additionalProperties: { type: string } }
        requiredFields: { type: object, additionalProperties: { type: string } }
    FileObject:
      type: object
      required: [fileId, status, mimeType, sizeBytes, sha256]
      properties:
        fileId: { type: string, format: uuid }
        status: { type: string, enum: [PENDING, READY, QUARANTINED, DELETED] }
        mimeType: { type: string }
        sizeBytes: { type: integer, minimum: 0 }
        sha256: { type: string }
...
    FaqConversationCreated:
      type: object
      additionalProperties: false
      required: [conversationId, status, startedAt]
      properties:
        conversationId: { type: string, format: uuid }
        status: { type: string, const: OPEN }
        startedAt: { type: string, format: date-time }
    AskFaqRequest:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message: { type: string, minLength: 2, maxLength: 2000 }
    FaqSource:
      type: object
      additionalProperties: false
      required: [position, title, section, question, category, url]
      properties:
        position: { type: integer, minimum: 1 }
        title: { type: string }
        section: { type: string }
        question: { type: [string, 'null'] }
        category: { type: [string, 'null'] }
        url: { type: [string, 'null'], format: uri }
    FaqFeedback:
      type: object
      additionalProperties: false
      required: [resolved, reasonCode]
      properties:
        resolved: { type: boolean }
        reasonCode: { type: [string, 'null'] }
    FaqMessage:
      type: object
      additionalProperties: false
      required: [id, role, body, faqHit, reasonCode, sources, feedback, createdAt]
      properties:
        id: { type: string, format: uuid }
        role: { type: string, enum: [TEACHER, ASSISTANT, SYSTEM] }
        body: { type: string }
        faqHit: { type: boolean }
        reasonCode:
          type: [string, 'null']
          enum: [FAQ_MATCHED, FAQ_CLARIFICATION_NEEDED, FAQ_NOT_FOUND, FAQ_AI_UNAVAILABLE, FAQ_AI_RESPONSE_INVALID, null]
        sources:
          type: array
          items: { $ref: '#/components/schemas/FaqSource' }
        feedback:
          oneOf:
            - { $ref: '#/components/schemas/FaqFeedback' }
            - { type: 'null' }
        createdAt: { type: string, format: date-time }
...
    FaqConversation:
      type: object
      additionalProperties: false
      required: [conversationId, status, startedAt, messages]
      properties:
        conversationId: { type: string, format: uuid }
        status: { type: string, enum: [OPEN, CLOSED, ARCHIVED] }
        startedAt: { type: string, format: date-time }
        messages:
          type: array
          items: { $ref: '#/components/schemas/FaqMessage' }
    FaqAnswerResponse:
      type: object
      additionalProperties: false
      required: [conversationId, teacherMessage, answer]
      properties:
        conversationId: { type: string, format: uuid }
        teacherMessage: { $ref: '#/components/schemas/FaqMessage' }
        answer: { $ref: '#/components/schemas/FaqMessage' }
    FaqFeedbackRequest:
      type: object
      additionalProperties: false
      required: [resolved]
      properties:
        resolved: { type: boolean }
        reasonCode: { type: string, minLength: 1, maxLength: 128 }
    FaqFeedbackResponse:
      type: object
      additionalProperties: false
      required: [messageId, resolved, reasonCode]
      properties:
        messageId: { type: string, format: uuid }
        resolved: { type: boolean }
        reasonCode: { type: [string, 'null'] }
    CreateAppEvent:
      type: object
      additionalProperties: false
      required:
        [eventName, eventId, eventSchemaVersion, sessionId, properties, occurredAt]
      properties:
        eventName: { $ref: '#/components/schemas/ClientAppEventName' }
        eventId: { type: string, minLength: 8, maxLength: 128 }
        eventSchemaVersion: { type: integer, const: 1 }
        sessionId: { type: string, minLength: 8, maxLength: 128 }
        taskAssignmentId: { type: string, minLength: 1, maxLength: 255 }
        properties:
          type: object
          maxProperties: 48
          additionalProperties:
            type: [string, number, boolean, 'null']
        occurredAt: { type: string, format: date-time }
    ClientAppEventName:
      type: string
      enum:
        - PAGE_VIEWED
        - PAGE_LOAD_SUCCEEDED
        - PAGE_LOAD_FAILED
        - NAVIGATION_CLICKED
        - LANGUAGE_CHANGED
        - PAGE_EXITED
        - TASK_CARD_IMPRESSION
        - TASK_CARD_CLICKED
        - TASK_OPENED_FROM_MESSAGE
        - TASK_DETAIL_VIEWED
        - TASK_STEP_STARTED
        - TASK_EXITED
        - TASK_PROGRESS_RESUMED
        - VIDEO_PLAYED
        - VIDEO_PAUSED
        - VIDEO_RESUMED
        - VIDEO_SEEK_BLOCKED
        - VIDEO_STALLED
        - VIDEO_FAILED
        - QUIZ_UNLOCKED
        - QUIZ_STARTED
        - QUIZ_SUBMITTED
        - QUIZ_PASSED
        - QUIZ_FAILED
        - DOCUMENT_READING_STARTED
        - DOCUMENT_MODULE_EXPANDED
        - DOCUMENT_COMPLETED
        - CHECKLIST_STARTED
        - CHECKLIST_PROGRESS_UPDATED
        - CHECKLIST_SUBMITTED
        - CAMERA_PERMISSION_RESULT
        - CAMERA_OPENED
        - CAMERA_PHOTO_TAKEN
        - CAMERA_PHOTO_RETAKEN
        - CAMERA_FAILED
        - UPLOAD_STARTED
        - UPLOAD_SUCCEEDED
        - UPLOAD_FAILED
        - UPLOAD_RETRIED
        - FAQ_OPENED
        - FAQ_QUESTION_SUBMITTED
        - FAQ_FEEDBACK_SUBMITTED
        - AI_HELP_OPENED
        - HELP_ROUTE_SELECTED
        - SUPPORT_TICKET_SUBMITTED
        - SUPPORT_TICKET_CREATED
        - SUPPORT_TICKET_OPENED
        - SUPPORT_TICKET_REPLY_READ
        - SUPPORT_TICKET_RESOLUTION_CONFIRMED
        - SUPPORT_TICKET_FOLLOWUP_SUBMITTED
        - MESSAGE_IMPRESSION
        - MESSAGE_DETAIL_OPENED
        - MESSAGE_TASK_CLICKED
        - MESSAGE_ACTION_EXPIRED
        - API_FAILED
        - NETWORK_TIMEOUT
        - FRONTEND_UNCAUGHT_ERROR
    AppEventAccepted:
      type: object
      additionalProperties: false
      required: [accepted, eventId]
      properties:
        accepted: { type: boolean, const: true }
        eventId: { type: string }
    ApiError:
      type: object
      required: [code, message, requestId, retryable]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        retryable: { type: boolean }
        details: { type: object, additionalProperties: true }